Added the code into the correct file, and removed it from urls.py
